сделано вычисление массы и цены с учетом bitPackag и vidPayment, Common переименованы в OrderCommon и CartCommon

+++ domain/src/main/java/com/lakomka/dto/ProductDto.java
@@ -15,7 +15,7 @@ public class ProductDto {
     private String article;
     private String unit;
     private String unitVid;
-    private Integer packag;
+    private Double packag;
     private BigDecimal price;
     private Integer weight;
     private Integer quantity;
